auth/requirements.txt
astroid==2.9.3
click==8.0.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.31
mysqlclient==2.1.0
parso==0.8.3
platformdirs==2.4.1
PyJWT==2.3.0
pylint==2.12.2
toml==0.10.2
Werkzeug==2.0.2
wrapt==1.13.3